move your posts to more indicative place & use BasicCard for rendering
