+++ ui/src/index.js
@@ -19,7 +19,7 @@ import PrivateRoutes from './utils/PrivateRoutes'
 import Login from './pages/Login/Login';
 import Profile from './pages/Profile';
 import Stream from './pages/Stream/Stream';
-import Posts from './pages/Posts';
+import YourPosts from './pages/YourPosts/YourPosts';
 import PostDetail from './components/PostDetail';
 import NotFound from './components/NotFound';
 import CreatePost from './pages/CreatePost';
@@ -35,7 +35,7 @@ root.render(
                <Route path='/' element={<App />}>
                   <Route element={<PrivateRoutes />}>
                      <Route path="/" element={<Stream/>} exact/>
-                     <Route path="/posts" element={<Posts />} />
+                     <Route path="/posts" element={<YourPosts />} />
                      <Route path="/posts/:postid" element={<PostDetail />} />
                      <Route path="/profile" element={<Profile />} />
                      <Route path="/createpost" element={<CreatePost />} />
