+++ ui/src/pages/Stream/Stream.js
@@ -53,15 +53,15 @@ const Stream = () => {
                                 <PostHeader 
                                     author={item.author} 
                                     title={item.title} 
-                                    time={(item.updated_at ? item.updated_at : item.publisehd)} />}
+                                    time={(item.updated_at ? item.updated_at : item.publisehd)} 
+                                />}
                             content={
                                 <PostContent 
                                     description={item.description}
                                     content={item.content}
-                                />
-                            }
+                                />}
                         />
-                    )
+                    );
                     break;
                 case 'comment':
                     // TODO render
